modify api struct to match geoloraserver

+++ models/model_gateway_reception_tdoa.go
@@ -2,15 +2,10 @@ package models
 
 type GatewayReceptionTdoa struct {
 	GatewayId string `json:"gatewayId"`
-
-	AntennaId int32 `json:"antennaId"`
-
-	Rssi float64 `json:"rssi"`
-
+	AntennaId int `json:"antennaId"`
+	Rssi int `json:"rssi"`
 	Snr float64 `json:"snr"`
-
-	Toa int32 `json:"toa"`
-
+	Toa int `json:"toa"`
 	EncryptedToa string `json:"encryptedToa"`
 
 	AntennaLocation *AntennaLocation `json:"antennaLocation,omitempty"`
